docs(responsive-record-view): implement R2 panel triage as single-pass rewrite
The R2 panel returned 3/3 REVISE across three lenses. Every finding was
accepted; none was rejected. This commit implements the full triage in
one rewrite rather than another incremental patch, because R2a traced
fourteen consistency defects to exactly that editing style.

Measurement corrections (verified against Vaadin 25.2.6 source before
acting). The "hard ceiling of 3 columns at 375px" claim was false: 100px
is Vaadin's default column width, not a minimum, so the relaxed-width
experiment measured the default back at itself. The harness is rewritten
with faithful flex-grow transcription and the Details cell corrected to
the real renderer's widest line, which lowers the flagship overflow
figures without changing the verdict: seven of ten views still overflow.
A new experiment measures the cheap width fix honestly: forcing the
journey view's four columns into 375px truncates about 54 percent of the
identity field. The refutation of the cheap fix now rests on measured
truncation rather than a false floor. The harness is renamed to
*.measure.ts so the CI glob never runs it.

Design corrections. The panel found that hiding all wide columns
disconnects their sorters, silently dropping sort from the dataProvider
query including declarative defaults. Section 3.4's bar is now specified
as header-hosted so sorters stay connected, with a dataProvider-level
fallback, an SC10 continuity assertion, and first place in the
implementation sequence. The row-template contract gains the slots the
pilot actually needs: actions as an overflow menu reusing the
grid-menu-a11y pattern, a sort surface for declaratively-sorted views, a
view-level controls region for the live connection and refresh toggles
that lived in column headers, and a chip driven by the item model rather
than the part-name mechanism no pilot view uses. Slot optionality and
the full pilot mapping resolve U-8a in the document. The artifact shape
is decided: template helper, per-view row-template module, and a
ReactiveController replacing ResponsiveMixin in migrated views. The
768px threshold is pinned with a derivation from the worst measured
content demand.

Consistency sweep. All fourteen R2a findings fixed: stale U-11 and
descriptor-model references removed, the accessibility-inheritance
contradiction corrected, SC1 and SC3 rekeyed to real artifacts, SC4 and
SC11 mapped in section 6, the wrong pilot figure corrected to one grid
with three columns, and the review trail extended.

Next step: a final R3 verification round, then user approval.
